Restructure asyncremote files
