+++ src/asyncremote/index.ts
@@ -1,3 +1,3 @@
-export * from './ipc';
+export * from './types';
 export * from './main';
 export * from './renderer';
